+++ src/js/components/parallax.js
@@ -37,7 +37,8 @@
         defaults: {
             velocity : 0.5,
             target   : false,
-            viewport : false
+            viewport : false,
+            media    : false
         },
 
         boot: function() {
@@ -95,7 +96,7 @@
             this.props    = {};
             this.velocity = (this.options.velocity || 1);
 
-            var reserved  = ['target','velocity','viewport','plugins'];
+            var reserved  = ['target','velocity','viewport','plugins','media'];
 
             Object.keys(this.options).forEach(function(prop){
 
@@ -128,6 +129,22 @@
         },
 
         process: function() {
+
+            if (this.options.media) {
+
+                switch(typeof(this.options.media)) {
+                    case 'number':
+                        if (window.innerWidth < this.options.media) {
+                            return false;
+                        }
+                        break;
+                    case 'string':
+                        if (window.matchMedia && !window.matchMedia(this.options.media).matches) {
+                            return false;
+                        }
+                        break;
+                }
+            }
 
             var percent = this.percentageInViewport();