+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Added uk-active class to form-select component
   - Added uk-flex-nowrap class to flex component
   - Added kenburnsanimations option to slideshow component
+  - Added media option to parallax component
   - Added activeClass option to tooltip component
   - Updated tooltip component to look for an updated title attribute before show
   - Updated slideset filter activation